requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
PyMuPDF